requirements.txt
APScheduler==3.9.1
arabic-reshaper==2.1.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-apscheduler==0.6.2
django-extensions==3.0.9
django-nested-admin==3.4.0
django-post-office==3.6.1
djangorestframework==3.14.0
future==0.18.2
html2text==2020.1.16
html5lib==1.0.1
httplib2==0.18.0
idna==3.4
jsonfield==3.1.0
lxml==4.5.2
Markdown==3.3.1
MonthDelta==0.9.1
mysqlclient==2.1.1
Pillow==7.2.0
pyfcm==1.5.4
PyPDF2==1.26.0
python-bidi==0.4.2
python-monkey-business==1.0.0
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
reportlab==3.5.42
requests==2.28.1
six==1.14.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
tinycss2==1.1.1
tomd==0.1.3
typing==3.7.4.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
webencodings==0.5.1
xhtml2pdf==0.2.5
